examples/adk-tracing/agents/research.py
"""Research agent that performs document retrieval and analysis."""
import asyncio
import random
from google import genai
from opentelemetry import trace, context
from opentelemetry.trace import Status, StatusCode
from .tracing_utils import get_tracer

class ResearchAgent:
    """Agent specialized in research and document retrieval."""
    
    def __init__(self):
        # Use Google AI Studio instead of Vertex AI for easier setup
        self.client = genai.Client()
        self.model_name = "gemini-2.5-flash"
        self.tracer = get_tracer()
    
    async def process(self, query: str, parent_context):
        """Process a research query with document retrieval."""
        with self.tracer.start_as_current_span(
            "research_agent.process",
            context=parent_context,
            attributes={
                "agent.type": "research",
                "query.text": query,
                "model.name": self.model_name
            }
        ) as span:
            try:
                # Step 1: Retrieve relevant documents
                documents = await self.retrieve_documents(query)
                
                # Step 2: Analyze the content
                analysis = await self.analyze_content(query, documents)
                
                # Step 3: Format results
                result = await self.format_results(analysis)
                
                span.set_status(Status(StatusCode.OK))
                span.set_attribute("documents.retrieved", len(documents))
                span.set_attribute("result.length", len(result))
                
                return result
                
            except Exception as e:
                span.record_exception(e)
                span.set_status(Status(StatusCode.ERROR, str(e)))
                raise
    
    async def retrieve_documents(self, query: str):
        """Simulate document retrieval from a knowledge base."""
        with self.tracer.start_span("retrieve_documents") as span:
            span.set_attribute("retrieval.query", query)
            
            # Simulate retrieval delay
            await asyncio.sleep(random.uniform(0.1, 0.3))
            
            # Mock document retrieval based on query
            documents = []
            
            if "quantum computing" in query.lower():
                documents = [
                    {
                        "title": "Quantum Computing Fundamentals",
                        "content": "Quantum computing leverages quantum mechanics principles like superposition and entanglement. Recent developments include improved qubit stability and error correction algorithms.",
                        "source": "IEEE Quantum Computing Journal, 2024"
                    },
                    {
                        "title": "Applications of Quantum Computing",
                        "content": "Key applications include cryptography, drug discovery, financial modeling, and optimization problems. IBM and Google have demonstrated quantum supremacy in specific tasks.",
                        "source": "Nature Quantum Information, 2024"
                    }
                ]
            elif "renewable energy" in query.lower():
                documents = [
                    {
                        "title": "Energy Storage Technologies",
                        "content": "Battery technology has advanced significantly with lithium-ion improvements and solid-state batteries. Grid-scale storage solutions now include pumped hydro and compressed air systems.",
                        "source": "Renewable Energy Review, 2024"
                    },
                    {
                        "title": "Recent Breakthroughs in Solar",
                        "content": "Perovskite solar cells have achieved 26% efficiency in labs. Bifacial panels and floating solar farms are expanding deployment options.",
                        "source": "Solar Energy Materials, 2024"
                    }
                ]
            elif "artificial general intelligence" in query.lower() or "agi" in query.lower():
                documents = [
                    {
                        "title": "AGI Safety Research",
                        "content": "Current AGI safety research focuses on alignment problems, interpretability, and robustness. Key areas include value learning, corrigibility, and safe exploration.",
                        "source": "AI Safety Conference Proceedings, 2024"
                    },
                    {
                        "title": "Progress Toward AGI",
                        "content": "While narrow AI excels in specific domains, AGI remains theoretical. Challenges include common sense reasoning, transfer learning, and consciousness.",
                        "source": "Journal of Artificial Intelligence Research, 2024"
                    }
                ]
            else:
                # Generic research documents
                documents = [
                    {
                        "title": f"Research on {query}",
                        "content": f"Recent developments in {query} show promising results. Multiple research teams are exploring various approaches.",
                        "source": "General Science Review, 2024"
                    }
                ]
            
            span.set_attribute("documents.count", len(documents))
            span.add_event("Documents retrieved", {"count": len(documents)})
            
            return documents
    
    async def analyze_content(self, query: str, documents):
        """Analyze retrieved documents using the LLM."""
        with self.tracer.start_span("analyze_content") as span:
            span.set_attribute("documents.input", len(documents))
            
            # Format documents for analysis
            doc_text = "\n\n".join([
                f"Title: {doc['title']}\nContent: {doc['content']}\nSource: {doc['source']}"
                for doc in documents
            ])
            
            # For demo purposes, we'll skip the actual LLM call
            # In production: model = self.client.models.get(self.model_name)
            
            prompt = f"""Analyze the following documents and provide a comprehensive response to the query: "{query}"

Documents:
{doc_text}

Provide a detailed analysis that:
1. Synthesizes information from the documents
2. Identifies key findings and developments
3. Notes any important caveats or limitations
4. Cites sources appropriately"""
            
            # Simulate LLM processing
            await asyncio.sleep(random.uniform(0.2, 0.4))
            
            # For demo purposes, create a mock response
            # In production, you would call: response = await model.generate_content_async(prompt)
            analysis = {
                "findings": [f"Key finding from research on {query}", "Recent developments show progress"],
                "sources_used": [doc['source'] for doc in documents],
                "confidence": 0.85
            }
            
            span.set_attribute("analysis.confidence", analysis['confidence'])
            span.add_event("Analysis completed")
            
            return analysis
    
    async def format_results(self, analysis):
        """Format the analysis into a readable result."""
        with self.tracer.start_span("format_results") as span:
            # Simulate formatting delay
            await asyncio.sleep(random.uniform(0.05, 0.1))
            
            # Create formatted output
            result = f"""Research Summary:

Key Findings:
{chr(10).join(f'• {finding}' for finding in analysis['findings'])}

Sources Consulted:
{chr(10).join(f'- {source}' for source in analysis['sources_used'])}

Confidence Level: {analysis['confidence']*100:.0f}%
"""
            
            span.set_attribute("format.type", "summary")
            span.set_attribute("result.length", len(result))
            
            return result 